Separate GeometrizeModelBase from GeometrizeModelHillClimb
